fix: change fill type from int to double

+++ nn/include/linalg.h
@@ -22,7 +22,7 @@ Matrix* read_matrix(char* fileName);
 Matrix* create_matrix(int rows, int cols);
 Matrix* copy_matrix(const Matrix* m);
 Matrix* matrix_flatten(Matrix* m); // Provided for intuitive code writing
-void fill_matrix(Matrix* m, int n);
+void fill_matrix(Matrix* m, double n);
 void randomize_matrix(Matrix* m);
 void free_matrix(Matrix* m);
 void write_matrix(Matrix* m, char* filename);
